Move to MEVIS to v1 and cleanup imports
