+++ src/unicorn_eval/adaptors/segmentation/mevis_conv_segmentation_3d/v1/__init__.py
@@ -1,0 +1,6 @@
+from unicorn_eval.adaptors.segmentation.mevis_conv_segmentation_3d.v1.main import \
+    ConvSegmentation3D
+
+__all__ = [
+    "ConvSegmentation3D",
+]
